requirements.txt
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning>=1.5.0
albumentations>=1.0.0
hydra-core>=1.1.0
omegaconf>=2.1.0
wandb>=0.12.0
pillow>=8.0.0
numpy>=1.20.0
tqdm>=4.62.0
